website/requirements.txt
# https://pip.pypa.io/en/latest/reference/requirements-file-format/
Django>=4.2,<5.1
wagtail>=6.0,<6.1
django-trim>=0.2
